+++ CHANGELOG.md
@@ -13,6 +13,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   path in chroot
 - Chroot to a PowerPC64 glibc world; tested on [Arch Linux Power]
 - Chroot to a [Debian] mips64el world; tested on [Debian MIPS Port]
+- Add the `ido(1)` and `ish(1)` option `--libdir` set `/lib:/usr/lib` as
+  default library path in chroot
 
 ### Removed